Adds and uses disableGen Type directive

+++ packages/graphback/src/resolvers/knex/targetResolverContext.ts
@@ -27,7 +27,7 @@ const knex = new KnexResolver()
  */
 
 const createResolver = (t: Type, database: string): string => {
-  if(t.config.create) {
+  if(t.config.create && !t.config.disableGen) {
     return knex.createTemplate(database, t.config.subCreate, getFieldName(t.name, ResolverType.CREATE), getTableName(t.name), t.name)
   }
 
@@ -35,7 +35,7 @@ const createResolver = (t: Type, database: string): string => {
 }
 
 const updateResolver = (t: Type): string => {
-  if(t.config.update) {
+  if(t.config.update && !t.config.disableGen) {
     return knex.updateTemplate(t.config.subUpdate, getFieldName(t.name, ResolverType.UPDATE), getTableName(t.name), t.name)
   }
 
@@ -43,7 +43,7 @@ const updateResolver = (t: Type): string => {
 }
 
 const deleteResolver = (t: Type): string => {
-  if(t.config.delete) {
+  if(t.config.delete && !t.config.disableGen) {
     return knex.deleteTemplate(t.config.subDelete, getFieldName(t.name, ResolverType.DELETE), getTableName(t.name), t.name)
   }
 
@@ -51,7 +51,7 @@ const deleteResolver = (t: Type): string => {
 }
 
 const findResolver = (t: Type): string => {
-  if(t.config.find) {
+  if(t.config.find && !t.config.disableGen) {
     return knex.findTemplate(getFieldName(t.name, ResolverType.FIND, 's'), getTableName(t.name))
   }
 
@@ -59,7 +59,7 @@ const findResolver = (t: Type): string => {
 }
 
 const findAllResolver = (t: Type): string => {
-  if(t.config.findAll) {
+  if(t.config.findAll && !t.config.disableGen) {
     return knex.findAllTemplate(getFieldName(t.name, ResolverType.FIND_ALL, 's'), getTableName(t.name))
   }
 
@@ -71,7 +71,7 @@ const findAllResolver = (t: Type): string => {
  */
 
 const newSub = (t: Type): string => {
-  if(t.config.create && t.config.subCreate) {
+  if(t.config.create && t.config.subCreate && !t.config.disableGen) {
     return knex.newSub(t.name)
   }
 
@@ -79,7 +79,7 @@ const newSub = (t: Type): string => {
 }
 
 const updatedSub = (t: Type): string => {
-  if(t.config.update && t.config.subUpdate) {
+  if(t.config.update && t.config.subUpdate && !t.config.disableGen) {
     return knex.updatedSub(t.name)
   }
 
@@ -87,7 +87,7 @@ const updatedSub = (t: Type): string => {
 }
 
 const deletedSub = (t: Type): string => {
-  if(t.config.delete && t.config.subDelete) {
+  if(t.config.delete && t.config.subDelete && !t.config.disableGen) {
     return knex.deletedSub(t.name)
   }
 
@@ -100,7 +100,7 @@ const deletedSub = (t: Type): string => {
  */
 const createSubscriptionTypes = (t: Type): string => {
   const subscriptionEnum = []
-  if(t.config.create && t.config.subCreate) {
+  if(t.config.create && t.config.subCreate && !t.config.disableGen) {
     subscriptionEnum.push(`NEW_${t.name.toUpperCase()} = 'new${t.name.toLowerCase()}'`)
   }
   if(t.config.update && t.config.subUpdate) {
